Move `get_empty_ds_with_time` to `seaice_ecdr.ancillary`.
This removes the `gridid_to_xr_dataarray` module
