+++ seaice_ecdr/initial_daily_ecdr.py
@@ -38,6 +38,7 @@ from pm_tb_data.fetch.nsidc_0001 import NSIDC_0001_SATS
 from seaice_ecdr._types import ECDR_SUPPORTED_RESOLUTIONS, SUPPORTED_SAT
 from seaice_ecdr.ancillary import (
     get_adj123_field,
+    get_empty_ds_with_time,
     get_invalid_ice_mask,
     get_land90_conc_field,
     get_land_mask,
@@ -46,7 +47,6 @@ from seaice_ecdr.ancillary import (
 from seaice_ecdr.cli.util import datetime_to_date
 from seaice_ecdr.constants import STANDARD_BASE_OUTPUT_DIR
 from seaice_ecdr.grid_id import get_grid_id
-from seaice_ecdr.gridid_to_xr_dataarray import get_empty_ds_with_time
 from seaice_ecdr.platforms import get_platform_by_date
 from seaice_ecdr.regrid_25to12 import reproject_ideds_25to12
 from seaice_ecdr.tb_data import EXPECTED_ECDR_TB_NAMES, EcdrTbData, get_ecdr_tb_data
